remove Query suffix from queries
