+++ src/Adapter/SqlManager/QueryHandler/GetSqlRequestForEditingHandler.php
@@ -29,7 +29,7 @@ namespace PrestaShop\PrestaShop\Adapter\SqlManager\QueryHandler;
 use PrestaShop\PrestaShop\Core\Domain\SqlManagement\EditableSqlRequest;
 use PrestaShop\PrestaShop\Core\Domain\SqlManagement\Exception\SqlRequestException;
 use PrestaShop\PrestaShop\Core\Domain\SqlManagement\Exception\SqlRequestNotFoundException;
-use PrestaShop\PrestaShop\Core\Domain\SqlManagement\Query\GetSqlRequestForEditingQuery;
+use PrestaShop\PrestaShop\Core\Domain\SqlManagement\Query\GetSqlRequestForEditing;
 use PrestaShop\PrestaShop\Core\Domain\SqlManagement\QueryHandler\GetSqlRequestForEditingHandlerInterface;
 use PrestaShop\PrestaShop\Core\Domain\SqlManagement\ValueObject\SqlRequestId;
 use RequestSql;
@@ -47,7 +47,7 @@ final class GetSqlRequestForEditingHandler implements GetSqlRequestForEditingHan
      * @throws SqlRequestException
      * @throws SqlRequestNotFoundException
      */
-    public function handle(GetSqlRequestForEditingQuery $query)
+    public function handle(GetSqlRequestForEditing $query)
     {
         $entity = $this->loadById($query->getRequestSqlId());
